Restore dummy files
