+++ practical/ProcessMining/__init__.py
@@ -1,3 +1,4 @@
+from .process_mining_dummy import hello_world
 from .kmeans import Kmeans
 
-__all__ = ["Kmeans"]
+__all__ = ["hello_world", "Kmeans"]
