+++ practical/ProcessMining/process_mining_dummy.py
@@ -1,0 +1,2 @@
+def hello_world(string: str) -> str:
+    return "Hello " + string
